+++ GoogleCloudExtension/GoogleCloudExtension/SolutionUtils/VsProject.cs
@@ -13,7 +13,10 @@
 // limitations under the License.
 
 using EnvDTE;
+using System;
 using System.IO;
+using System.Linq;
+using System.Xml.Linq;
 
 namespace GoogleCloudExtension.SolutionUtils
 {
@@ -22,7 +25,13 @@ namespace GoogleCloudExtension.SolutionUtils
     /// </summary>
     internal class VsProject : ISolutionProject
     {
+        private const string MsbuildNamespace = "http://schemas.microsoft.com/developer/msbuild/2003";
+        private const string WebApplicationGuid = "{349c5851-65df-11da-9384-00065b846f21}";
+
         private readonly Project _project;
+        private readonly Lazy<KnownProjectTypes> _knownProjectType;
+
+        #region ISolutionProject
 
         public string DirectoryPath => Path.GetDirectoryName(_project.FullName);
 
@@ -30,11 +39,36 @@ namespace GoogleCloudExtension.SolutionUtils
 
         public string Name => _project.Name;
 
-        public KnownProjectTypes ProjectType => _project.GetProjectType();
+        public KnownProjectTypes ProjectType => _knownProjectType.Value;
+
+        #endregion
 
         public VsProject(Project project)
         {
             _project = project;
+            _knownProjectType = new Lazy<KnownProjectTypes>(GetProjectType);
+        }
+
+        private KnownProjectTypes GetProjectType()
+        {
+            var dom = XDocument.Load(_project.FullName);
+            var projectGuids = dom.Root
+                .Elements(XName.Get("PropertyGroup", MsbuildNamespace))
+                .Descendants(XName.Get("ProjectTypeGuids", MsbuildNamespace))
+                .Select(x => x.Value)
+                .FirstOrDefault();
+
+            if (projectGuids == null)
+            {
+                return KnownProjectTypes.None;
+            }
+
+            var guids = projectGuids.Split(';');
+            if (guids.Contains(WebApplicationGuid))
+            {
+                return KnownProjectTypes.WebApplication;
+            }
+            return KnownProjectTypes.None;
         }
     }
 }
